Machine_learning/requirements.txt
fastapi==0.96.0
tensorflow==2.12.0
uvicorn==0.22.0
Pillow
numpy
python-multipart==0.0.6